refactor: restructure pkg/ for better separation of concerns
- Split pkg/parse and pkg/types into focused packages:
  - pkg/link: Pure data structures (AnchorTag, SitemapUrl, results)
  - pkg/fetcher: HTTP operations (GetHTML)
  - pkg/parser: HTML/XML parsing (html.go, sitemap.go)
  - pkg/resolver: URL resolution logic (ResolveAbsoluteUrl, IsSpecialLink)

- Remove GetHttpStatus() methods from types (separation of concerns)
- Replace single-letter import aliases with descriptive names
- Migrate all tests to new structure (87-95% coverage maintained)
- Add .github/copilot-instructions.md for AI agent guidance

Benefits:
- Clear separation of concerns for better maintainability
- Easier to test and mock individual components
- Flexible foundation for concurrent HTTP checking
- Ready for future features (checker, formatter packages)

All tests passing. No breaking changes to CLI interface.
