+++ README.md
@@ -7,30 +7,37 @@ A tool for parsing and analyzing HTML links.
 ### Running Tests
 
 Run all unit tests:
+
 ```bash
 go test ./...
 ```
 
 Run tests with coverage:
+
 ```bash
-go test -v -race -coverprofile=coverage.out ./...
+go test -v -race -coverprofile=coverage.out ./...\
 ```
 
 ### Code Quality
 
 #### Formatting
+
 Format all Go code:
+
 ```bash
 go fmt ./...
 ```
 
 Check formatting without making changes:
+
 ```bash
 gofmt -s -l .
 ```
 
 #### Linting
+
 Lint code using golangci-lint:
+
 ```bash
 golangci-lint run ./...
 ```
@@ -48,6 +55,7 @@ The workflow configuration can be found in `.github/workflows/pr-checks.yml`.
 ### Building
 
 Build the project:
+
 ```bash
 go build .
 ```
@@ -57,16 +65,19 @@ This will create a `brokli` binary in the current directory.
 ### Usage
 
 Run the CLI tool:
+
 ```bash
 ./brokli
 ```
 
 Check a URL:
+
 ```bash
 ./brokli check url https://example.com
 ```
 
 Check a sitemap:
+
 ```bash
 ./brokli check sitemap https://example.com/sitemap.xml
 ```
